requirements/base.txt
requests>=2.3.0
Jinja2>=2.7.3
locustio==0.7.2
pyzmq==14.3